requirements.txt
streamlit
langchain
chromadb
google-generative-ai
python-dotenv